chore: RainbowRecyclerAssetList
